+++ www/modules/email/controller/AccountController.php
@@ -267,7 +267,7 @@ class AccountController extends \GO\Base\Controller\AbstractModelController
 			$stmt =  \GO\Email\Model\Account::model()->find($findParams);
 
 			while ($account = $stmt->fetch()) {
-				$account->maybeRefreshToken();
+				$account->maybeRenewAccessToken();
 				$alias = $account->getDefaultAlias();
 				if($alias){
 					$nodeId=base64_encode('account_' . $account->id);