+++ CHANGELOG.md
@@ -1742,6 +1742,7 @@
 - Core: Fixed date range with commentedAt filter
 - docuwiki: correct path in module description to INSTALL.TXT
 - ActiveSync: Was generating unnecessary temporary files
+- OAuth2Client / Email: Refactor renewing of access token, better check access token on sending email
 
 16-11-2020 6.4.192
 - Comments: Fixed bug when adding two comments without opening another item would overwrite the first.